ai-api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
ollama>=0.4.0
python-dotenv>=1.0.0
